--- webService/src/main/java/com/asklepios/hospitalreservation_asklepios/Service/IF_LikeService.java
@@ -1,4 +1,0 @@
-package com.asklepios.hospitalreservation_asklepios.Service;
-
-public interface IF_LikeService {
-}
