+++ webService/src/main/java/com/asklepios/hospitalreservation_asklepios/Service/IM_BoardService.java
@@ -2,6 +2,7 @@ package com.asklepios.hospitalreservation_asklepios.Service;
 
 import com.asklepios.hospitalreservation_asklepios.Repository.IF_BoardMapper;
 import com.asklepios.hospitalreservation_asklepios.VO.BoardVO;
+import com.asklepios.hospitalreservation_asklepios.VO.LikeVO;
 import com.asklepios.hospitalreservation_asklepios.VO.PageVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -124,4 +125,29 @@ public class IM_BoardService implements IF_BoardService{
         boardMapper.deleteBoard(no);
     }
 
+    public boolean checkLike(LikeVO likeVO){
+        boolean flag =false;
+        if(boardMapper.checkLike(likeVO)!=0) {
+            flag = true;
+            boardMapper.minusLikeCount(likeVO);
+            boardMapper.delLike(likeVO);
+
+        }else{
+            boardMapper.plusLikeCount(likeVO);
+            boardMapper.addLike(likeVO);
+        }
+        return flag;
+    }
+    public boolean firstLike(LikeVO likeVO){
+        int check=boardMapper.checkLike(likeVO);
+        boolean flag =false;
+        if(check==1) {
+            flag=true;
+        }
+        return flag;
+    }
+
+    public int countHeart(LikeVO likeVO){
+        return boardMapper.selectHeart(likeVO);
+    }
 }
